jenkins

+++ Jenkinsfile
@@ -4,7 +4,7 @@ pipeline {
         DATABASE_URI = credentials("DATABASE_URI")
         AUTHOR = credentials("AUTHOR")
         app_version=1
-        rollback='true'
+        rollback='false'
     }
     stages{
         stage('Test'){
@@ -14,10 +14,12 @@ pipeline {
         }
         stage("Build"){
             steps{
-                sh "docker-compose build --parallel --build-arg APP_VERSION=${app_version} && docker-compose push"
-                sh "docker system prune -af"
-                sh "bash jenkins/build_images.sh"
-            }  
+                if (env.rollback == 'false'){
+                    sh "docker-compose build --parallel --build-arg APP_VERSION=${app_version} && docker-compose push"
+                    sh "docker system prune -af"
+                    sh "bash jenkins/build_images.sh"
+                }   
+            }
         }
         stage('Configure ansible'){
             steps {
